Pulling various test mock methods and classes out to testMock file of their own
